fusion enttmngr with gamescene
