+++ Carom/src/SceneSystem/GameScene.h
@@ -1,16 +1,19 @@
-#ifndef GAME_Scene_H
-#define GAME_Scene_H
+#pragma once
 
+#include <SDLUtils.h>
 #include <list>
-
+#include <vector>
+#include <array>
+#include <box2d/box2d.h>
 #include "GameList.h"
+#include "ecs.h"
+
+#include "Entity.h"
 
 class Game;
 
 // Declaraciones anticipadas
 namespace ecs{
-	class Entity;
-	class EntityManager;
 
 	/**
  * Estado abstracto del juego.
@@ -21,15 +24,81 @@ namespace ecs{
  */
 class GameScene
 {
-	GameList<ecs::Entity> entities;
-	EntityManager* _enttmngr; // TODO: decidir ciclo de los objectos
+	GameList<Entity> _entities;
+	std::array<std::vector<entity_t>, maxGroupId> _entsByGroup;
 
 protected:
 	Game* game;
 
 	GameScene(Game* game);
 
-	// void addObject(ecs::Entity* obj); // TODO: decidir ciclo de los objectos
+	// ! fusion EM
+
+	void createWhiteBall(Vector2D pos, b2BodyType type, float density, float friction, float restitution, float radius); // TODO
+
+	void createEffectBall(effect::effectId effectId, Vector2D pos, b2BodyType type, float density, float friction, float restitution, float radius); // TODO
+
+	void createTable(/* type */); // TODO
+
+	inline GameList<Entity>& getEntities() { return _entities; }
+
+	// Setting the state of the entity (alive or dead)
+	//
+	inline void setAlive(entity_t e, bool alive) {
+		e->setAlive(alive);
+	}
+
+	// Returns the state of the entity (alive o dead)
+	//
+	inline bool isAlive(entity_t e){
+		return e->isAlive();
+	}
+
+	// Adds a component to an entity. It receives the type T (to be created),
+	// and the list of arguments (if any) to be passed to the constructor.
+	//
+	template<typename T, typename ...Ts>
+	inline void addComponent(entity_t e, Ts &&... args) {
+		// the component id
+		static_assert(cmpId<T> < ecs::maxComponentId);
+
+		// create, initialise and install the new component
+		T *c = new T(e, std::forward<Ts>(args)...);
+
+		if (!e->addComponent<T>(c)) {
+			delete c;
+		}
+	}
+
+	// Removes the component T, if any, from the entity.
+	template<typename T>
+	inline bool removeComponent(entity_t e) {
+		return e->removeComponent<T>();
+	}
+
+	// returns the vector of all entities
+	//
+	inline auto& getEntities(grpId_t gId) {
+		return _entsByGroup[gId];
+	}
+
+	// return true if there is a component with identifier T::id in the entity
+	//
+	template<typename T>
+	inline bool hasComponent(entity_t e) {
+		return e->tryGetComponent(cmpId<T>);
+	}
+
+	// Returns the component, of the entity, that corresponds to position T,
+	// casting it to T*. The casting is done just for ease of use, to avoid casting
+	// outside.
+	//
+	template<typename T>
+	inline T* getComponent(entity_t e) {
+		return e->getComponent<T>(cmpId<T>);
+	}
+
+	// ! fusion
 
 public:
 	virtual ~GameScene();
@@ -51,8 +120,3 @@ GameScene::getGame() const
 }
 
 }
-
-
-
-
-#endif // GAME_Scene_H
